backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-phonenumber-field==7.3.0
django-rest-knox==4.2.0
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
phonenumbers==8.13.32
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
